Add files via upload
EXP 5-a - Sobel Operator = The Sobel operator is a well-known edge detection method used in image processing. It detects the edges in an image by calculating the gradient of the image intensity at each pixel. The operator uses convolution with two distinct kernels to detect horizontal and vertical edges. The Sobel operator helps you do just that by scanning the image and highlighting areas where the brightness changes significantly. It's like a special filter that emphasizes the edges, making them stand out. The Sobel operator uses a small window, typically 3x3 pixels, to examine the image. It calculates the difference in brightness between neighboring pixels in the window, giving more importance to pixels closer to the center. When it finds an area with a significant change in brightness, it marks that area as an edge. The resulting image shows the edges highlighted, making it easier to see the shapes and boundaries within the original picture.

 EXP 5-b - Previtt operator = The Prewitt operator is another edge detection tool used in image processing. It works similarly to the Sobel operator, but with a few key differences. Instead of using a weighted difference, the Prewitt operator calculates a simple difference between neighboring pixels in the window. This means it treats all pixels in the window equally, without giving more importance to those closer to the center. Similar to the Sobel operator, the Prewitt operator uses convolution with specific kernels to detect edges. The Prewitt operator is better suited for detecting edges in multiple directions, rather than just horizontal and vertical. It's also less sensitive to noise, which can be beneficial when working with images that have a lot of random variations in brightness.